refactor: restructure conversion tracking by moving properties and types to dedicated interfaces for improved organization and type safety
